flat_field needs to support different data models
When getting the value of the DISPAXIS keyword for NIRSpec spectroscopic data,
the flat_field step used the syntax that would work for an ImageModel (or
similar), but sometimes the input is a MultiSlitModel.  For the latter case,
the keyword will be gotten from the first slit in the model.

See #3767.