flat_field needs to support different data models (#3807)
* Don't use pixels with wavelength = NaN

The master_background/expand_to_2d.get_wavelengths() function is now used
to get the 2-D array of wavelengths from the input file, instead of explicitly
checking for the existence of the wavelength attribute (an extension in the
input file), and evaluating the WCS function if the attribute does not exist
or is not populated.

The replace_bad_values function has been modified.  Instead of setting the
science data to a fill value (zero) for pixels where the data value was NaN or
the data quality flag included DO_NOT_USE, the science data are now set to NaN
wherever the wavelength is NaN or the DQ flag includes DO_NOT_USE.

Closes #3223.

* Update the change log.

* Populate and use a new DISPAXIS keyword

Function get_dispersion_direction (lib/dispaxis.py) was modified to check
that some of the function arguments are not None before trying to convert
them to upper case.

Function load_wcs (assign_wcs/assign_wcs.py) was modified to get a value
and populate the DISPAXIS keyword in wcsinfo, for spectroscopic data.

The schemas for wcsinfo and multispec (datamodels/schemas/) were modified
to include a dispersion_direction attribute for keyword DISPAXIS.

The extract_1d step now uses the dispersion_direction attribute, replacing
the find_dispaxis function.  The attribute is also copied to the output file.

Functions extract_tso_object and extract_grism_objects (extract_2d/grisms.py)
were modified to copy the dispersion_direction attribute from input to output.

The flat_field step now uses the dispersion_direction attribute, replacing
the find_dispaxis function.  The calling sequences of a number of functions
were modified, either to add dispaxis or to change the location of dispaxis
in the calling sequence.

Functions do_drizzle and find_dispersion_axis (resample/resample_spec.py)
were modified; find_dispersion_axis uses meta.wcsinfo.dispersion_direction
to get the dispersion direction, and do_drizzle includes 'dispersion_direction'
in the list of attributes that are copied from input to output.

Closes #3767

* Update change log.

* Import warnings in resample_spec.  Delete a test in flat_field.

* Fixed unit test for find_dispersion_axis.

* Further changes for DISPAXIS

In extract_2d, keyword DISPAXIS is also copied from input to output for
NIRSpec data.
Documentation for the affected steps has been modified to mention the
DISPAXIS keyword.

* Update change log

* flat_field needs to support different data models

When getting the value of the DISPAXIS keyword for NIRSpec spectroscopic data,
the flat_field step used the syntax that would work for an ImageModel (or
similar), but sometimes the input is a MultiSlitModel.  For the latter case,
the keyword will be gotten from the first slit in the model.

See #3767.

* Update the change log