JP-878 Don't try to compute pixel solid angle for WFSS data
The extract_1d step was modified to only call util.pixel_area() for the
first integration of multi-integration data.  Also, that function immediately
returns pixel_solid_angle = None for WFSS data; for these data the WCS
function returns constant values for right ascension and declination, so
the angular size of a pixel cannot be determined by this function.

The extract_2d step was modified to compute and populate the wavelength
attribute for NIRCam TSO data.  This partially addresses JP-845, GitHub #3784,
Compute wavelengths for NIRCam TSO data, but only the wavelength attribute
is populated, not the variance arrays.  To close this issue, the variance
arrays need to be copied from input to output.