Modify the photom step to use new data models
The photom step has been modified to use new photom data models.  This
version is incompatible with the older photom data models.  The changes
include:

For spectroscopic data, the data are multiplied by the 2D relative response
term; previously the data were divided by this term.  The relevant code
changes for this were copied from PR #4033.

For NIRSpec spectroscopic data for an extended target, the conversion factor
will be divided by the pixel area in steradians, because the value in the new
photom reference file converts to MJy, but we need MJy / sr.  This will also
be done for NIRISS SOSS data if the target is extended, but that case should
never be encountered.

The pixel area is gotten from the area reference file and copied to metadata.
This was done before but only for imaging data; now it is also done for
spectroscopic data.  Another difference is that the pixel area is only gotten
from the area reference file; pixel area keywords in the photom reference
file are no longer checked.