JP-1053 Fix bugs in photom
Function save_area_nirspec failed for NIRSpec IFU data because it was trying
to assign a value to slit.meta.photometry.pixelarea_arcsecsq, but when running
photom the input is a single 2048x2048 array, i.e. there are no slits.

Function photom_io failed for NIRISS extended-source data.  The conversion
factor should be divided by the pixel area, but that value was None.