JP-1106 Modify extract_1d because of changes to photom
The input file may have one set of units or another, depending on the
instrument and whether the target is a point source or extended; this is
due to recent changes to the photom step.  As a result, the way the table
columns are computed by the extract_1d step depends on the configuration.

There's now a check for whether the photom step has been run, and if it has
not been, the output from extract_1d will be in units of DN/s.

The pixel solid angle is not gotten from a header keyword, rather than
being computed from the WCS.

The source type (point or extended) is gotten from either the primary header
or an extension header, depending on the input data model.

Closes  JP-1106, GitHub #4207 and JP-859, GitHub #3477
These tickets can be closed:  JP-999, GitHub #4032 and JP-1005, GitHub #4040