JP-1152 Check whether the output pixel number is NaN
For each input spectrum, the array of corresponding pixel numbers in the
output, combined spectrum is computed by using the inverse WCS function.
It is possible for some of these pixel numbers (probably only at an endpoint
of the input spectrum) to be NaN due to the finite bounding box for the WCS.
A check for NaN was added.

While testing this change, another bug was uncovered.  In the section that
tests for the possibility that there was no input data for some output pixel,
the error estimate for flux was being assigned to the error estimate for
surface brightness.  This has been fixed.