JP-1225 MIRI masterbackground tests copied to new regtest
Tests of the master_background step for MIRI data have been created in the
new regression test system, using input data copied from the old regtests.
There is a test with a user-supplied background file, a test with each of
LRS and MRS data for the case of nodded exposures, and a similar pair of tests
for the case of dedicated background exposures.