JP-1232 Populate the background column
For the case that the target is a point source and the instrument is NIRSpec
or NIRISS (SOSS), the input units are megajanskys, and the background was
being set to zero.  The extracted background is now divided by the pixel area
(solid angle) to convert to units of surface brightness.