Modifications to extract_2d for grism data
For NIRCam TSO GRISM data, the variance arrays and the INT_TIMES table are
now copied from input to output.  This was overlooked in previous updates.
The related issues are JP-845 (GitHub #3784) to copy the variance arrays,
and JP-1249 (GitHub #4489) to copy the INT_TIMES table.

For NIS_WFSS, NRC_WFSS, or NRC_TSGRISM data, the SLTSTRT1 and SLTSTRT2 keywords
now give the start location of the 2-D cutout in the input image.  Previously,
these keywords were 1.  The related issue is JP-903 (GitHub #3872).