Fix for Issue #4571, Jump Stepping Flagging too many pixels
This change has the fix for the problem described in #4571. The number of frames per group is now passed to the two_pointdifference routine rather than the number of groups. The bug led to large underestimates of the read noise and thus many false postive jumps. This also caused the step to run very slowly.
  A new test was added to recreate the problem. It failed before the fix and now passes.