JP-954 Swap alpha and beta arrays; add support for a DQ extension
The alpha and beta arrays from the IRS2Model reference file were used
incorrectly.  The alpha arrays are now used for the reference output, and
the beta arrays are now used for the interleaved reference pixels.

If there is a DQ array in the reference file, this will now be used to
set some of the reference pixel values to zero.