JP-1311 Fix bugs; add unit tests
Function `decode_mask` was sometimes returning an ndarray instead of a list.
The readout direction in `decode_mask` was incorrect.
The readout direction in the code that is now in `clobber_ref` was not taken
into account for DQ table column `odd_even`.
A unit test for `clobber_ref` and one for `decode_mask` by itself were added.