JP-1154: Flag CR neighbors  #4297 (#4434)
* Improve Variable Names and Comments

In preperation for the code walkthrough I updated the comments and renamed a few variables to be more descriptive.
  I also updated a couple of the test cases.

* added ignore for profile information and anything starting with “.”

* Update .gitignore

* Create twopoint_difference.py

* CR Neighbor flagging

Initial Changes to add flagging neihbors of CRs

* Changes mainly for new parameters

The changes add three new config parameters for the flagging of CR neighbors and pass them to the detect_jumps method

* add code to return jump ratios

The all_ratios cube is now returned. It is needed to find the thresholds for CR neighbor flagging in detect_jumps.

* Add flagging of CR neighbors

This routine needed to handle the flagging of neighbors because the neigbhors could cross slice boundaries if find_crs is running on multiple processes.

* Tests for flagging of neighbors

New tests for flagging neighbors when requested were added. Also, calling parameters to detect_jump were updated for older tests.

* Don't track any pycharm environment changes

* add tests for multiple ints and slices

This adds a multi-int test and one that makes sure that CRs at the top and bottom of a slice are correctly flagged.

* Updated calling parameters to detect_jumps

One test was not updated earlier

* Move-flagging-neighbors-to-twopoint

The flagging of neighbors was moved to twopoint_difference to allow it to be run on multiple processes.

* Added Comments for changes

* Remove Pycharm file

* Remove pycharm file

* Remove Pycharm file

* Remove pycharm file

* remove pycharm file

* Remove Pycharm file

* Remove pycharm file

* remove pycharm file

* remove pycharm file

* Update test_detect_jumps.py

Added doc strings to the tests and added more asserts to the tests to make sure the routine is working correctly.

* Update test_twopoint_difference.py

Changed tests to use new calling parameters and returned values from find_crs.

* Updated jump parameters

Added the four jump parameters that were missing.  "max_cores" was missing from the last update. Also, needed for the flag-CR-neighbors was flag_4_neighbors, max_jump_to_flag_nieghbors and min_jump_to_flag_neighbors

* Update jump_step.py

Fixed the typo in the name of the flag_4_neighbors parameter

* Update jump_step.py

Removed test for do_y_intercept, it can't be reached and is not planned for implementation

* Delete yintercept.py

The routine is not used and is not in current future plans. If needed, it could be retrieved from an earlier release.

* Update jump.py

Remove references to the y-intercept method.

* delete blank line

* Add tests for subarray and bad gain pixels

Two tests were added to complete code coverage: A test that has bad gain values in the reference file and a test for a full frame reference file being used in subarray.

* Update test_jump_step.py

A test to show that when the integration has 2 groups the step is skipped and is marked 'SKIPPED' in the model.

* Removed unreachable code

Removed assert statements that are never reached in a test that is expected to fail. There is not point to have code that can't be executed.

* Delete all_11_radon.txt

Should not be in repo

* Delete all_radon.txt

should not be in repo

* Delete all_raw_radon.txt

should not be in repo

* Remove Pycharm files for local .gitignore

* Changes for RampModel and skip test for CR in every pixel

* Changes based on Comments from Phil Hodge

These are mostly clean up issues. I changed on of the tests for jump at slice boundaries to only run when the number of slices is greater than 1. Otherwise, the test would write outside of the boundaries of the array.

* Changes based on comment from Howard Bushouse

I have removed all references to the median slopes except interal to twopoint_difference where it is used to find the jumps.
I also removed any reference to the y-intercept method.

* Fix for Issue #4571, Jump Stepping Flagging too many pixels

This change has the fix for the problem described in #4571. The number of frames per group is now passed to the two_pointdifference routine rather than the number of groups. The bug led to large underestimates of the read noise and thus many false postive jumps. This also caused the step to run very slowly.
  A new test was added to recreate the problem. It failed before the fix and now passes.

* Improve config parameter comments

The units of the various thresholds were update to show that they are in units of sigma.

* Update return slice comments

Update the comments to correctly reflect what is returned for each slice

* Remove new parameters

I removed the new parameters to deal with the problem that a parameter cannot be set to "None"

* Change cube size to save memory

For the 100 group test the number of rows and number of columns was reduced by a factor of 10 to save memory during the TravisCI testing.

* Flake 8 updates

Remove unused variable and line that Flake8 claims is always True (It is not always True but is redundant with the remaining assert.

* Removed unused variables

Removing variables that Flake8 flagged as unused.

* remove unused variable

Fix a Flake8 flagged problem