ramp fit updates
Moved ols_ramp_fit_multi back into this file. Removed print statements.
Changed the power calculation for optimal fitting to be interpolated to
not have discontinuities in the slopes.
Changed variable to be be double precision float to avoid and rounding
errors. Also, removed some of the zeroing out of arrays to allow for
debugging.