Ramp multiprocessing (#4815)
* Improve Variable Names and Comments

In preperation for the code walkthrough I updated the comments and renamed a few variables to be more descriptive.
  I also updated a couple of the test cases.

* added ignore for profile information and anything starting with “.”

* Update .gitignore

* Create twopoint_difference.py

* CR Neighbor flagging

Initial Changes to add flagging neihbors of CRs

* remove conflicting changes

Partial changes on wrong branch removed.

* update jump.py

Try to fix conflict

* Remove Pycharm file

* Remove Pycharm File

* Delete workspace.xml

* Delete vcs.xml

* Delete other.xml

* Delete modules.xml

* Delete misc.xml

* Delete Project_Default.xml

* Delete codeStyleConfig.xml

* Delete coverage file

* Delete coverage file

* Delete coverage file

* Update step.py

* Initial changes to add multi-proc to ramp fit

The first draft of changes to make ramp_fit able to run with multiprocessing.

* Switch to explict parameter passing

changing to not use RampModel

* Removed RampModel as input variable

This allows starmap to work. The multiprocessing now works but has the same pickle problem for the returned models.

* update return values from ols_ramp_fit

Update the return values to return the calculaed arrays rather than the models due to the pickle problem.

* Update ramp_fit.py

* fix output arrays

Changes to get the output arrays to have the correct indexing

* Moved back into ramp_fit.py

* Fixes for two utils problems

The special case of zero cosmic rays caused the optional output cr_mag
to have a zero dimension.
Also, the scaling of the read noise was changing the input array. That
lead to issues with multiple calls. Making a copy of the array fixed
the problem.

* ramp fit updates

Moved ols_ramp_fit_multi back into this file. Removed print statements.
Changed the power calculation for optimal fitting to be interpolated to
not have discontinuities in the slopes.
Changed variable to be be double precision float to avoid and rounding
errors. Also, removed some of the zeroing out of arrays to allow for
debugging.

* Adding multiprocessing parameter

Changes to pass multiprocessing parameter to ols_ramp_filt_multi

* New multiprocessing test

Added to to compare results of single processor to multiprocessing for
ramp fit.

* removed commented out code

* update to master

* catch up to master

* Add multiprocessing to generalized least squares fitting method

* update thresholds to match double precision results

* update calling parameters to ramp_fit

* add in gls_output_integ method

* update to master

* Remove conflict labels and changes

* Update ramp_fit.py

change log.info to log.debug to match master

* Update jump.py

Put pool.starmap back in instead pool.map

* Update jump.py

Add Comment

* Refix ramp_fit.py

Again fixed calling parameters to ramp_fit_multi

* remove print

* added back CR

* removed commented out prints

* added comment

* skip multi-proc tests

See if this fixes the Travis memory problems

* Update test_ramp_fit_cases.py

* Go back to f32 in ramp fit

* Update test_ramp_fit.py

fixed pytest dectorator

* Update ramp_fit.py

add pool.close and pool.join to see if this helps the Travis CI problem

* Don't use multiprocessing for single slices

* restore multi-proc tests

put back multiprocessing tests

* add close() and join after starmap to clear up processes

* skip instead of xfail GLS test

* add close and join for GLS

* remove GLS tests from paramaterize

* Don't do Multi-proc for GLS when  max_cores ==1

* Update ramp_fit.py

fix syntax

* update doc strings and import

* fix flake8 issues

* removed unused int_times

* Create description.rst

* add back files

* get current versions

* add back references_general

* add back resample

* Changes based on comments

* updated GLS - OLS comparision of multi-int

* Update test_ramp_fit.py

removed comment from skip for GLS to OLS multi-int test

* Update ramp_fit.py

changed to use variables as indices in slices of input and output arrays.

* add spaces to comments in jump.py

* Remove duplicate line

Removed duplicate line in ramp_fit