Add build_nirspec_output_wcs() method
This method will handle NIRSpec FS and MSA data for resample_spec
where there is subpixel dithering in the spectral direction.  It
Computes an output frame where the RA/Dec gets mapped only to
position in the slit, ignoring that dithers in the dispertion
direcitons are really mapping different parts of the sky.