median filtering handles bad pixels slightly better
For MIRI only, now if a pixel with a bad pixel flag (e.g DO_NOT_USE) is outside
the science aperture, it is simply set to 0 instead of median filtered. Median
filtering the unilluminated regions of the subarray threw over ten thousand
warnings, causing the alignment algorithm to take hours to days, instead of a
minute or less.