calwebb_coron3.align_refs's median filter now skips NON_SCIENCE
Now uses the DQ flags to sort the bad_locations. NON_SCIENCE pixels are
clustered, so the median filter would set them to 0. Now we just set them to 0
right away and avoid 10,000+ warning messages that get logged and cause the
runtime to be 1000x longer than it needs to be.