JP-2515: Workaround for excessive align_refs logging for MIRI coronagraphy (#6722)
* replaced np.median with np.nanmedian in align_refs

Since the pipeline is worried about NaNs ruining the computation of the median
value, it seems better to used nanmedian, which filters out NaNs when computing
the median.

* added print statements for debugging

* removed debug statement from align_refs_step

* median filtering handles bad pixels slightly better

For MIRI only, now if a pixel with a bad pixel flag (e.g DO_NOT_USE) is outside
the science aperture, it is simply set to 0 instead of median filtered. Median
filtering the unilluminated regions of the subarray threw over ten thousand
warnings, causing the alignment algorithm to take hours to days, instead of a
minute or less.

* Added median filter logging back in

* calwebb_coron3.align_refs's median filter now skips NON_SCIENCE

Now uses the DQ flags to sort the bad_locations. NON_SCIENCE pixels are
clustered, so the median filter would set them to 0. Now we just set them to 0
right away and avoid 10,000+ warning messages that get logged and cause the
runtime to be 1000x longer than it needs to be.

* Applied some fixes

- reinstated a log warning that I'd commented out by accident
- Fixed docstring style for consistency
- Minor change to NON_SCIENCE flag finder for speed
- added a few comments and changed variable names for readability

* Reinstated check for MIRI instrument

The NON_SCIENCE filter now only runs on MIRI data. There is a check for MIRI
(but not specifically for MIRI coronagraphy). Also removed unused code and
expanded comments for clarity.

* added align_refs changes to change log