Some testing for ZEROFRAME.
Adding ZEROFRAME processing to superbias.

Adding ZEROFRAME functionality to refpix, as well as adding keyword checking as a way to start ZEROFRAME processing in linearity and saturation steps.

Making style changes and properly check for the ZEROFRAME keyword for superbias processing.

Adding test for using ZEROFRAME for saturated ramps.

Adding PR line to the change log for ramp fitting.

Making changes due to code review.

Adding ZEROFRAME processing to IRS2 and adding a test for it.

Updating comments for the ZEROFRAME test of IRS2 saturation.

Adding new testing.