Ensuring no path data in JSON files for 'expnames'.
Adding test for path information in expnames in JSON association file.

Expanding documentation for input directories and directions for editing an association file.

Expanding documentation for input file requiremen ts and the input directory.

Fixing a failed test.

Updating the change log.

Changing raising an exception to raising a warning.