remove duplicate enum values from core.schema.yaml
jsonschema draft 4 (used by asdf) stats that enum values
MUST be unique. This was unchecked prior to a recent commit
python-jsonschema/jsonschema@2db9c58

core.schema.yaml contains duplicate dithering types for the
different instruments (for example CYCLING for NIRSpec IFU
and MIRI imaging)

This commit removes the duplicates in a way that might impair
user readability of the file (as types are no longer logically
grouped to instrument).