Add test for duplicate data when resaving model to fits
When resaving a model to fits, asdf loses track of array-hdu
links and saves arrays in both hdu and as internal blocks in the
asdf tree. This can make an asdf tree that is larger than what
can be stored in a fits hdu.

See asdf issue: https://github.com/asdf-format/asdf/issues/1232

This test is skipped as it fails with the current asdf version.

modify imports and use context manager for test_fits

Co-authored-by: James Davies <jdavies@mpia.de>

use context manager in test_fits

Co-authored-by: James Davies <jdavies@mpia.de>

improve test docstring in test_fits