implement new rule Asn_Lv2SpecImprint
This rule is needed to treat the imprint exposures and science.
The Asn_Lv2SpecSpecial should have done that. However, imprints
can be auxiliary to the already auxiliary exposures. Hence, once
an imprint is consumed in Asn_Lv2SpecSpecial, it is not available
to create its own Asn_Lv2SpecSpecial. Hence, a dedicated rule is
needed.