JP-3079 Ensure all imprint exposures are include in associations (#7438)
* reprocess imprints

this solves the issue.

however, will do some refactoring to condense and eliminate some
repetitive constraints.

* combine background and imprint comparison into single constraint

* create the generic imprint constraint

* factor out the background constraint

* apply refactor background constraint to other rules

* change order of precedence for exposure type determination

* fix return of the special override for exposure determination

* Include imprints in background and other special level2 processing

* do not add imprints to an imprint association

* implement new rule Asn_Lv2SpecImprint

This rule is needed to treat the imprint exposures and science.
The Asn_Lv2SpecSpecial should have done that. However, imprints
can be auxiliary to the already auxiliary exposures. Hence, once
an imprint is consumed in Asn_Lv2SpecSpecial, it is not available
to create its own Asn_Lv2SpecSpecial. Hence, a dedicated rule is
needed.

* update all unit and regressions

* fix logging of the command line arguments

* remove unnecessary list and tuple definitions from constraint values

* update changelog

---------

Co-authored-by: Jonathan Eisenhamer <eisenham@mentant.local.stsci.edu>