re-add quick check for alternative engineering database test
The test_environmental test when not run on the vpn
will fail to contact the alternative engineering database
server. This results in the test taking ~10 minutes waiting
for a response before turning into a skip.

This commit adds some lines of code that were removed in
PR #3022 which changed the skip from a decorator to a
pytest.skip when the exception/timeout occurs. This should
allow the test to skip quickly when the host is not accessable
(and instead skips in <1 second). The pytest.skip call
during the except block is left to deal with cases when the
host resolves but the url fails to resolve.