First attempt to modify rules to include both S200A1 and S200A2.
Moving the constraint name to force the existence of an attribute.

Removing debugging comment and removing unneeded lines.

Updating the change log.

Updating spectrographic test for level 3 due to the update of optical path constraints.

Updating class constraint to properly set name values and updated classes and a get function to accomodate this.

Updating constraints so 'sources' is a list.

Style fix.