TST: Delete pool_015_spec_nirspec_lv2bkg_reversed from test suite.
Tyler Pauly: I believe this was intended as a check against
behavior change with order of pool rows - it should be
the same content as pool 010, but with the background exposures first.
Because this code is only ever exercised for pools that DMS generates,
I think it’s okay to ignore testing parameter space outside DMS bounds
- that is, flight pool order is set by DMS, and we have test coverage
of those. So we can probably just ignore pool 015.

[ci skip]